refactor(core): remove getConfigDir/getConfigPath forwarding wrappers from config.ts, standardize to paths.ts
Fixes #330
